A couple of mail server bug fixes: Disabled user is not added to mail server on startup. Deleting a disabled user no longer throws an exception.

+++ src/main/java/com/smockin/mockserver/engine/MockedMailServerEngine.java
@@ -196,7 +196,11 @@ public class MockedMailServerEngine {
     public void removeMailUser(final MailMock mailMock) {
         logger.debug("removeMailUser called");
 
-        final SmockinGreenMailUserWrapper user = findGreenMailUser(mailMock.getAddress());
+        final SmockinGreenMailUserWrapper user = mailUsersMap.get(mailMock.getAddress());
+
+        if (user == null) {
+            return;
+        }
 
         user.getUser().delete();
